+++ crates/miden-protocol/src/transaction/executed_tx.rs
@@ -208,12 +208,6 @@ pub struct TransactionMeasurements {
     pub tx_script_processing: usize,
     pub epilogue: usize,
     pub auth_procedure: usize,
-    /// The number of cycles the epilogue took to execute after compute_fee determined the cycle
-    /// count.
-    ///
-    /// This is used to get the total number of cycles the transaction takes for use in
-    /// compute_fee itself.
-    pub after_tx_cycles_obtained: usize,
 }
 
 impl TransactionMeasurements {
@@ -238,7 +232,6 @@ impl Serializable for TransactionMeasurements {
         self.tx_script_processing.write_into(target);
         self.epilogue.write_into(target);
         self.auth_procedure.write_into(target);
-        self.after_tx_cycles_obtained.write_into(target);
     }
 }
 
@@ -250,7 +243,6 @@ impl Deserializable for TransactionMeasurements {
         let tx_script_processing = usize::read_from(source)?;
         let epilogue = usize::read_from(source)?;
         let auth_procedure = usize::read_from(source)?;
-        let after_tx_cycles_obtained = usize::read_from(source)?;
 
         Ok(Self {
             prologue,
@@ -259,7 +251,6 @@ impl Deserializable for TransactionMeasurements {
             tx_script_processing,
             epilogue,
             auth_procedure,
-            after_tx_cycles_obtained,
         })
     }
 }